gcp/esame-23-febbraio-23/src/requirements.txt
flask
flask-restful
WTForms
requests
google-cloud-firestore
gunicorn
python-dateutil
google-cloud-pubsub